test(core): use uvu
